+++ src/components/Card/Card.jsx
@@ -7,6 +7,7 @@ import Card from 'react-bootstrap/Card';
 import ListGroup from 'react-bootstrap/ListGroup';
 import Skeleton from '@mui/material/Skeleton';
 import './Card.css'
+import {addToFav, addToCart} from './FavAndCart';
 
 export default function ProductCard({ id, name, price, rating, image, brand }) {
 
@@ -17,33 +18,6 @@ export default function ProductCard({ id, name, price, rating, image, brand }) {
       setLoading(false);
     }, 1500);
   }, [])
-
-  const addToFav = () => {
-    let favs = JSON.parse(localStorage.getItem('favList'))
-    if (!favs) {
-      favs = [{ id, name, price, rating, image, brand }]
-      localStorage.setItem('favList', JSON.stringify(favs))
-      return
-    }    
-    if (favs.length < 10 && favs.every(item => item.id !== id)) {
-      favs.push({ id, name, price, rating, image, brand })
-      localStorage.setItem('favList', JSON.stringify(favs))
-      return
-    }
-    alert("Can't add the item, is alredy in the FavList or you have 10 Favs.")
-  }
-  const addToCart = () => {
-    let cart = JSON.parse(localStorage.getItem('cartList'))
-    if (cart) {
-      if (cart.length <= 10 && cart.every(item => item.id !== id)) {
-        cart.push({ id, name, price, rating, image, brand })
-      }
-    } else {
-      cart = [{ id, name, price, rating, image, brand }]
-    }
-    localStorage.setItem('totalPrice', JSON.stringify(price + JSON.parse(localStorage.getItem('totalPrice'))))
-    localStorage.setItem('cartList', JSON.stringify(cart))
-  }
 
   return (
     <Card className="card" >
@@ -74,8 +48,8 @@ export default function ProductCard({ id, name, price, rating, image, brand }) {
         </ListGroup>
       </Card.Body>
       <div className='containerButton'>
-        <BsStarFill className='CardIcon' onClick={addToFav} />
-        <BsCartFill className='CardIcon' onClick={addToCart}/>
+        <BsStarFill className='CardIcon' onClick={() => addToFav(id, name, price, rating, image, brand)} />
+        <BsCartFill className='CardIcon' onClick={() => addToCart(id, name, price, rating, image, brand)}/>
       </div>
     </Card>
   )
